Refactor: Created generatePlayer method and renamed className 'empty' to 'wall'.
Generating the player's character is moved from getInitialStatus to the dedicated method generatePlayer.
This is done since player will be generated in each dungeon.
The player appears in red color in the grid.
Once the page is loaded the grid is scrolled to show the player.

empty class represent a wall, so its name is changed to wall.