Feat: Player can now move from dungeon 0 to dungeon 4.
Player can go to the stairs 'purple squares' and go to the next dungeon.
Can go up to dungeon 4.
Added incrementDungeon method in GameArea which is passed as props to Maze.
In Maze, the code in getInitialState is moved to a seperate method initializeGrid which is called from getInitialState.
initialzeGrid generates a grid and loads the player, enemies, and all other items.
This initializeGrid method is called from presshandler when the player reaches a stairs cell.