Bug: Fixed bug where weapon don't change when dungeon changes.
The way the game works that in each dungeon a different weapon is found.
The bug was that if the player don't pick the knife in dungeon 0, the weapon in dungeon 4 will still be kninfe instead of RBG.
This is now fixed and each dungeon has a different weapon.
To acheive this, a new seletedWeapon propoerty is added to the state in GameArea,
weaponCounter is initialized to default value of 1 instead of 0,
weaponCounter is only incremented when the dungeon is incremented.