Feat: Each enemy is now given random health within a range.
When each enemy is created, it's given health value.
The value depends on the dungeon number and falls within a range.,
For example, in dungeon 0 enemies will heave health ranging from 10 to 15. Dungeon 1 20 to 25 and so on.
Modifications are done in generateEnemies method, also an array enemies containing positions and health is returned to initializeGrid and then to getInitialState.