Refactor & Feat: Refactored getInitialState in GameArea, Player now can attack and kill enemies.
Refactor: getInitialState in GameArea is refactored to help add functionality of attacking enemies.
weapons array in state has a new property damage which hold the damage value for each weapon.
Added a new property to the state 'weaponDamage'.
weaponCounter is abandoned. Weapon selection is based on the dungeon number.
Dungeon number now ranges from 1-5 instead of 0-4.
weaponDamage is passed as props to Maze.

Feat: Player can attack and kill enemies.
The health of each enemy is updated and saved after each attack.
An if statement is added to handlepess method to check if the new poisition has an enemy occupying it,
if so the new attackEnemy method is called which checks the health of the enemy.
UP TO THIS POINT, the enemy does not inflect damage on the player, this is the next step.