Feat: Player can now attack and kill boss.
The boss is given health which is stored in the state of Maze.
A new if statement is added to presshandler which checks if the new cell is has the class boss.
The method attackEnemy is modified a bit to allow for checking if the player won the game or not.
Next step is to implement XP and new levels and attack powers for the player.