Feat: XP and player level now functional.
As player kills enemies, the XP points to the next level updates. When the user collects all needed XP, level increases.
XP depend on dungeon number, the larger the dungeon number, the higher XP gained for killing an enemy, and the higher points needed to go to next level.
Once XP to level are all gained, level and health increases. Level increases by one. Health by multiplier of level number.
Added xp, xMultiplier and level propoerties to the state of GameArea.
Added updateXp method to GameArea which is passed to Maze.
updateXp is called from presshandler method if attackEnemy method returns true.
Next: numbers need to be tweaked to make sure the game is not ridiculously easy or imposibbly hard.