Feat: Numbers are tweaked and the game is now playable.
The game is not so easy to win nor impossibleto win.
The recommended action for player is find the weapon at each dungeon before killing any enemy.
Each enemy get hearder to beat at each new dungeon, without upgrading the weapon, it'll be tough to beat and the player will sustain a lot damage.