Refactor: Moved all audio varialbe declarations to the getInitialState method of Maze.
Before these variable were declared each time in presshandler method,
which resulted in some lag in playing audio because the varialbe needed to be declared and sound loaded then play the sound.
Now the variables and sound are declared and loaded once in the getInitialState,
only the sound is played in the presshandler method which removed the lag even with fast motion.