Feat: Instead of alert, the player receives a message after game is over with the option to start a new  game.
When the game is over, the message tells the user if it's a loss or win.
There is also a 'New Game' button to start a new game.
This message is found in the newly added component Message whose display is controlled by the state property messageStatus.
Properties added to the GameArea state are: messages 'contains default win and loss messages',
messageStatus which controls the display of Message component.
message which holds the message to be displayed,
gameProgress which is set to true as long as a game is running.
Two methods added to GameArea: displayMessage and handleClick which handles clicking the newGame button.
A new method newGrid is added to Maze and called from GameArea via refs. This method is called upon clicking the New Game button.
Other new changes include preventing keyboard from moving player when message is displayed.
Scrolling is locked as well when the meesage is displayed.