Reafctor & Feat: As the player moves, the visible tiles move as well.
Feat: The visible tiles move around with the player while moving through the dungeons. The player at the center of the visible tiles.
Refactor: To acheive this no new code was written.
Instead, setting tiles to dark was moved to playerVisibility method instead of setting the class directly during creating them.
This method is called from movePlayer method, so the visible tiles change when the player moves.
The method now checks if the dark property (in GameArea state and passed as props to Maze) is true,
if not, the method doesn't execute and all tiles would be visible.
Next: activate Turn lights on 'will be renamed to Toggle light'.